Add files via upload
Hello this is a pretty new project i'm still making it better.
I added a feature where it randomly changes speed to slow or fast to make it harder.
The speed depends on how fast your computer is. If your computer is slow the ball becomes slow, if your computer is fast the ball is fast.